+++ app/vmauth/main_test.go
@@ -4,10 +4,14 @@ import (
 	"bytes"
 	"fmt"
 	"io"
+	"net"
 	"net/http"
 	"net/http/httptest"
 	"strings"
+	"sync/atomic"
 	"testing"
+
+	"github.com/VictoriaMetrics/VictoriaMetrics/lib/netutil"
 )
 
 func TestRequestHandler(t *testing.T) {
@@ -40,6 +44,13 @@ func TestRequestHandler(t *testing.T) {
 			t.Fatalf("cannot initialize http request: %s", err)
 		}
 
+		r.RequestURI = r.URL.RequestURI()
+		r.RemoteAddr = "42.2.3.84:6789"
+		r.Header.Set("X-Forwarded-For", "12.34.56.78")
+		r.Header.Set("Connection", "Some-Header,Other-Header")
+		r.Header.Set("Some-Header", "foobar")
+		r.Header.Set("Pass-Header", "abc")
+
 		w := &fakeResponseWriter{}
 		if !requestHandler(w, r) {
 			t.Fatalf("unexpected false is returned from requestHandler")
@@ -57,49 +68,98 @@ func TestRequestHandler(t *testing.T) {
 	// regular url_prefix
 	cfgStr := `
 unauthorized_user:
-  url_prefix: {BACKEND}/foo?bar=baz
-`
+  url_prefix: {BACKEND}/foo?bar=baz`
 	requestURL := "http://some-host.com/abc/def?some_arg=some_value"
 	backendHandler := func(w http.ResponseWriter, r *http.Request) {
-		fmt.Fprintf(w, "requested_url=http://%s%s", r.Host, r.URL)
+		var bb bytes.Buffer
+		if err := r.Header.Write(&bb); err != nil {
+			panic(fmt.Errorf("unexpected error when marshaling headers: %w", err))
+		}
+		fmt.Fprintf(w, "requested_url=http://%s%s\n%s", r.Host, r.URL, bb.String())
 	}
 	responseExpected := `
 statusCode=200
 requested_url={BACKEND}/foo/abc/def?bar=baz&some_arg=some_value
-`
+Pass-Header: abc
+User-Agent: vmauth
+X-Forwarded-For: 12.34.56.78, 42.2.3.84`
 	f(cfgStr, requestURL, backendHandler, responseExpected)
 
 	// keep_original_host
 	cfgStr = `
 unauthorized_user:
   url_prefix: "{BACKEND}/foo?bar=baz"
-  keep_original_host: true
-`
+  keep_original_host: true`
 	requestURL = "http://some-host.com/abc/def"
 	backendHandler = func(w http.ResponseWriter, r *http.Request) {
 		fmt.Fprintf(w, "requested_url=http://%s%s", r.Host, r.URL)
 	}
 	responseExpected = `
 statusCode=200
-requested_url=http://some-host.com/foo/abc/def?bar=baz
-`
+requested_url=http://some-host.com/foo/abc/def?bar=baz`
 	f(cfgStr, requestURL, backendHandler, responseExpected)
 
-	// override request host
+	// override user-agent header
+	cfgStr = `
+unauthorized_user:
+  url_prefix: "{BACKEND}/foo?bar=baz"
+  headers:
+  - "User-Agent: foobar"`
+	requestURL = "http://some-host.com/abc/def"
+	backendHandler = func(w http.ResponseWriter, r *http.Request) {
+		fmt.Fprintf(w, "requested_url=http://%s%s\nUser-Agent=%s", r.Host, r.URL, r.Header.Get("User-Agent"))
+	}
+	responseExpected = `
+statusCode=200
+requested_url={BACKEND}/foo/abc/def?bar=baz
+User-Agent=foobar`
+	f(cfgStr, requestURL, backendHandler, responseExpected)
+
+	// delete user-agent header
+	cfgStr = `
+unauthorized_user:
+  url_prefix: "{BACKEND}/foo?bar=baz"
+  headers:
+  - "User-Agent:"`
+	requestURL = "http://some-host.com/abc/def"
+	backendHandler = func(w http.ResponseWriter, r *http.Request) {
+		fmt.Fprintf(w, "requested_url=http://%s%s\nUser-Agent=%s", r.Host, r.URL, r.Header.Get("User-Agent"))
+	}
+	responseExpected = `
+statusCode=200
+requested_url={BACKEND}/foo/abc/def?bar=baz
+User-Agent=Go-http-client/1.1`
+	f(cfgStr, requestURL, backendHandler, responseExpected)
+
+	// override request host with non-empty host
 	cfgStr = `
 unauthorized_user:
   url_prefix: "{BACKEND}/foo?bar=baz"
   headers:
   - "Host: other-host:12345"
-`
+  - "abc:"`
 	requestURL = "http://some-host.com/abc/def"
 	backendHandler = func(w http.ResponseWriter, r *http.Request) {
 		fmt.Fprintf(w, "requested_url=http://%s%s", r.Host, r.URL)
 	}
 	responseExpected = `
 statusCode=200
-requested_url=http://other-host:12345/foo/abc/def?bar=baz
-`
+requested_url=http://other-host:12345/foo/abc/def?bar=baz`
+	f(cfgStr, requestURL, backendHandler, responseExpected)
+
+	// override request host with empty host
+	cfgStr = `
+unauthorized_user:
+  url_prefix: "{BACKEND}/foo?bar=baz"
+  headers:
+  - "Host:"`
+	requestURL = "http://some-host.com/abc/def"
+	backendHandler = func(w http.ResponseWriter, r *http.Request) {
+		fmt.Fprintf(w, "requested_url=http://%s%s", r.Host, r.URL)
+	}
+	responseExpected = `
+statusCode=200
+requested_url={BACKEND}/foo/abc/def?bar=baz`
 	f(cfgStr, requestURL, backendHandler, responseExpected)
 
 	// /-/reload handler failure
@@ -109,16 +169,14 @@ requested_url=http://other-host:12345/foo/abc/def?bar=baz
 	}
 	cfgStr = `
 unauthorized_user:
-  url_prefix: "{BACKEND}/foo"
-`
+  url_prefix: "{BACKEND}/foo"`
 	requestURL = "http://some-host.com/-/reload"
 	backendHandler = func(_ http.ResponseWriter, _ *http.Request) {
 		panic(fmt.Errorf("backend handler shouldn't be called"))
 	}
 	responseExpected = `
 statusCode=401
-The provided authKey doesn't match -reloadAuthKey
-`
+The provided authKey doesn't match -reloadAuthKey`
 	f(cfgStr, requestURL, backendHandler, responseExpected)
 	if err := reloadAuthKey.Set(origAuthKey); err != nil {
 		t.Fatalf("unexpected error: %s", err)
@@ -128,8 +186,7 @@ The provided authKey doesn't match -reloadAuthKey
 	cfgStr = `
 users:
 - username: foo
-  url_prefix: "{BACKEND}/bar"
-`
+  url_prefix: "{BACKEND}/bar"`
 	requestURL = "http://some-host.com/a/b"
 	backendHandler = func(_ http.ResponseWriter, _ *http.Request) {
 		panic(fmt.Errorf("backend handler shouldn't be called"))
@@ -137,8 +194,7 @@ users:
 	responseExpected = `
 statusCode=401
 Www-Authenticate: Basic realm="Restricted"
-missing 'Authorization' request header
-`
+missing 'Authorization' request header`
 	f(cfgStr, requestURL, backendHandler, responseExpected)
 
 	// incorrect authorization
@@ -146,92 +202,99 @@ missing 'Authorization' request header
 users:
 - username: foo
   password: secret
-  url_prefix: "{BACKEND}/bar"
-`
+  url_prefix: "{BACKEND}/bar"`
 	requestURL = "http://foo:invalid-secret@some-host.com/a/b"
 	backendHandler = func(_ http.ResponseWriter, _ *http.Request) {
 		panic(fmt.Errorf("backend handler shouldn't be called"))
 	}
 	responseExpected = `
 statusCode=401
-Unauthorized
-`
+Unauthorized`
 	f(cfgStr, requestURL, backendHandler, responseExpected)
+
+	// incorrect authorization with logging invalid auth tokens
+	origLogInvalidAuthTokens := *logInvalidAuthTokens
+	*logInvalidAuthTokens = true
+	cfgStr = `
+users:
+- username: foo
+  password: secret
+  url_prefix: "{BACKEND}/bar"`
+	requestURL = "http://foo:invalid-secret@some-host.com/a/b?c=d"
+	backendHandler = func(_ http.ResponseWriter, _ *http.Request) {
+		panic(fmt.Errorf("backend handler shouldn't be called"))
+	}
+	responseExpected = `
+statusCode=401
+remoteAddr: "42.2.3.84:6789, X-Forwarded-For: 12.34.56.78"; requestURI: /a/b?c=d; cannot authorize request with auth tokens ["http_auth:Basic Zm9vOmludmFsaWQtc2VjcmV0"]`
+	f(cfgStr, requestURL, backendHandler, responseExpected)
+	*logInvalidAuthTokens = origLogInvalidAuthTokens
 
 	// correct authorization
 	cfgStr = `
 users:
 - username: foo
   password: secret
-  url_prefix: "{BACKEND}/bar"
-`
+  url_prefix: "{BACKEND}/bar"`
 	requestURL = "http://foo:secret@some-host.com/a/b"
 	backendHandler = func(w http.ResponseWriter, r *http.Request) {
 		fmt.Fprintf(w, "requested_url=http://%s%s", r.Host, r.URL)
 	}
 	responseExpected = `
 statusCode=200
-requested_url={BACKEND}/bar/a/b
-`
+requested_url={BACKEND}/bar/a/b`
 	f(cfgStr, requestURL, backendHandler, responseExpected)
 
 	// verify how path cleanup works
 	cfgStr = `
 unauthorized_user:
-  url_prefix: {BACKEND}/foo?bar=baz
-`
+  url_prefix: {BACKEND}/foo?bar=baz`
 	requestURL = "http://some-host.com/../../a//.///bar/"
 	backendHandler = func(w http.ResponseWriter, r *http.Request) {
 		fmt.Fprintf(w, "requested_url=http://%s%s", r.Host, r.URL)
 	}
 	responseExpected = `
 statusCode=200
-requested_url={BACKEND}/foo/a/bar/?bar=baz
-`
+requested_url={BACKEND}/foo/a/bar/?bar=baz`
 	f(cfgStr, requestURL, backendHandler, responseExpected)
 
 	// verify how path cleanup works for url without path
 	cfgStr = `
 unauthorized_user:
-  url_prefix: {BACKEND}/foo?bar=baz
-`
+  url_prefix: {BACKEND}/foo?bar=baz`
 	requestURL = "http://some-host.com/"
 	backendHandler = func(w http.ResponseWriter, r *http.Request) {
 		fmt.Fprintf(w, "requested_url=http://%s%s", r.Host, r.URL)
 	}
 	responseExpected = `
 statusCode=200
-requested_url={BACKEND}/foo?bar=baz
-`
+requested_url={BACKEND}/foo?bar=baz`
 	f(cfgStr, requestURL, backendHandler, responseExpected)
 
 	// verify how path cleanup works for url without path if url_prefix path ends with /
 	cfgStr = `
 unauthorized_user:
-  url_prefix: {BACKEND}/foo/?bar=baz
-`
+  url_prefix: {BACKEND}/foo/?bar=baz`
 	requestURL = "http://some-host.com/"
 	backendHandler = func(w http.ResponseWriter, r *http.Request) {
 		fmt.Fprintf(w, "requested_url=http://%s%s", r.Host, r.URL)
 	}
 	responseExpected = `
 statusCode=200
-requested_url={BACKEND}/foo/?bar=baz
-`
+requested_url={BACKEND}/foo/?bar=baz`
 	f(cfgStr, requestURL, backendHandler, responseExpected)
+
 	// verify how path cleanup works for url without path and the url_prefix without path prefix
 	cfgStr = `
 unauthorized_user:
-  url_prefix: {BACKEND}/?bar=baz
-`
+  url_prefix: {BACKEND}/?bar=baz`
 	requestURL = "http://some-host.com/"
 	backendHandler = func(w http.ResponseWriter, r *http.Request) {
 		fmt.Fprintf(w, "requested_url=http://%s%s", r.Host, r.URL)
 	}
 	responseExpected = `
 statusCode=200
-requested_url={BACKEND}/?bar=baz
-`
+requested_url={BACKEND}/?bar=baz`
 	f(cfgStr, requestURL, backendHandler, responseExpected)
 
 	// verify routing to default_url
@@ -240,17 +303,166 @@ unauthorized_user:
   url_map:
   - src_paths: ["/foo/.+"]
     url_prefix: {BACKEND}/x-foo/
-  default_url: {BACKEND}/404.html
-`
+  default_url: {BACKEND}/404.html`
 	requestURL = "http://some-host.com/abc?de=fg"
 	backendHandler = func(w http.ResponseWriter, r *http.Request) {
 		fmt.Fprintf(w, "requested_url=http://%s%s", r.Host, r.URL)
 	}
 	responseExpected = `
 statusCode=200
-requested_url={BACKEND}/404.html?request_path=http%3A%2F%2Fsome-host.com%2Fabc%3Fde%3Dfg
-`
+requested_url={BACKEND}/404.html?request_path=http%3A%2F%2Fsome-host.com%2Fabc%3Fde%3Dfg`
 	f(cfgStr, requestURL, backendHandler, responseExpected)
+
+	// verify routing to default url_prefix
+	cfgStr = `
+unauthorized_user:
+  url_map:
+  - src_paths: ["/foo/.+"]
+    url_prefix: {BACKEND}/x-foo/
+  url_prefix: {BACKEND}/default`
+	requestURL = "http://some-host.com/abc?de=fg"
+	backendHandler = func(w http.ResponseWriter, r *http.Request) {
+		fmt.Fprintf(w, "requested_url=http://%s%s", r.Host, r.URL)
+	}
+	responseExpected = `
+statusCode=200
+requested_url={BACKEND}/default/abc?de=fg`
+	f(cfgStr, requestURL, backendHandler, responseExpected)
+
+	// missing default_url and default url_prefix for unauthorized user
+	cfgStr = `
+unauthorized_user:
+  url_map:
+  - src_paths: ["/foo/.+"]
+    url_prefix: {BACKEND}/x-foo/`
+	requestURL = "http://some-host.com/abc?de=fg"
+	backendHandler = func(_ http.ResponseWriter, _ *http.Request) {
+		panic(fmt.Errorf("backend handler shouldn't be called"))
+	}
+	responseExpected = `
+statusCode=400
+remoteAddr: "42.2.3.84:6789, X-Forwarded-For: 12.34.56.78"; requestURI: /abc?de=fg; missing route for http://some-host.com/abc?de=fg`
+	f(cfgStr, requestURL, backendHandler, responseExpected)
+
+	// missing default_url and default url_prefix for unauthorized user when there are configs for authorized users
+	cfgStr = `
+users:
+- username: some-user
+  url_map:
+  - src_paths: ["/foo/.+"]
+    url_prefix: {BACKEND}/x-foo/
+unauthorized_user:
+  url_map:
+  - src_paths: ["/abc/.*"]
+    url_prefix: {BACKEND}/x-bar`
+	requestURL = "http://some-host.com/abc?de=fg"
+	backendHandler = func(_ http.ResponseWriter, _ *http.Request) {
+		panic(fmt.Errorf("backend handler shouldn't be called"))
+	}
+	responseExpected = `
+statusCode=401
+Www-Authenticate: Basic realm="Restricted"
+missing 'Authorization' request header`
+	f(cfgStr, requestURL, backendHandler, responseExpected)
+
+	// all the backend_urls are unavailable for unauthorized user
+	cfgStr = `
+unauthorized_user:
+  url_map:
+  - src_paths: ["/foo/.*"]
+    url_prefix:
+    - http://127.0.0.1:1/
+    - http://127.0.0.1:2/`
+	requestURL = "http://some-host.com/foo/?de=fg"
+	backendHandler = func(_ http.ResponseWriter, _ *http.Request) {
+		panic(fmt.Errorf("backend handler shouldn't be called"))
+	}
+	responseExpected = `
+statusCode=503
+remoteAddr: "42.2.3.84:6789, X-Forwarded-For: 12.34.56.78"; requestURI: /foo/?de=fg; all the 2 backends for the user "" are unavailable`
+	f(cfgStr, requestURL, backendHandler, responseExpected)
+
+	// all the backend_urls are unavailable for authorized user
+	cfgStr = `
+users:
+- username: some-user
+  url_map:
+  - src_paths: ["/foo/.*"]
+    url_prefix:
+    - http://127.0.0.1:1/
+    - http://127.0.0.1:2/`
+	requestURL = "http://some-user@some-host.com/foo/?de=fg"
+	backendHandler = func(_ http.ResponseWriter, _ *http.Request) {
+		panic(fmt.Errorf("backend handler shouldn't be called"))
+	}
+	responseExpected = `
+statusCode=503
+remoteAddr: "42.2.3.84:6789, X-Forwarded-For: 12.34.56.78"; requestURI: /foo/?de=fg; all the 2 backends for the user "some-user" are unavailable`
+	f(cfgStr, requestURL, backendHandler, responseExpected)
+
+	// zero discovered backend IPs
+	customResolver := &fakeResolver{
+		Resolver: &net.Resolver{},
+		lookupIPAddrResults: map[string][]net.IPAddr{
+			"some-addr": {},
+		},
+	}
+	origResolver := netutil.Resolver
+	netutil.Resolver = customResolver
+	cfgStr = `
+unauthorized_user:
+  url_prefix: ['http://some-addr:1234/foo/bar']
+  discover_backend_ips: true`
+	requestURL = "http://abc.com/def/?de=fg"
+	backendHandler = func(_ http.ResponseWriter, _ *http.Request) {
+		panic(fmt.Errorf("backend handler shouldn't be called"))
+	}
+	responseExpected = `
+statusCode=503
+remoteAddr: "42.2.3.84:6789, X-Forwarded-For: 12.34.56.78"; requestURI: /def/?de=fg; all the 0 backends for the user "" are unavailable`
+	f(cfgStr, requestURL, backendHandler, responseExpected)
+	netutil.Resolver = origResolver
+
+	// retry_status_codes failure
+	var retries atomic.Int64
+	cfgStr = `
+unauthorized_user:
+  url_prefix: ['{BACKEND}/path1', '{BACKEND}/path2']
+  retry_status_codes: [500, 502]`
+	requestURL = "http://some-host.com/foo/?de=fg"
+	backendHandler = func(w http.ResponseWriter, _ *http.Request) {
+		retries.Add(1)
+		w.WriteHeader(500)
+	}
+	responseExpected = `
+statusCode=503
+remoteAddr: "42.2.3.84:6789, X-Forwarded-For: 12.34.56.78"; requestURI: /foo/?de=fg; all the 2 backends for the user "" are unavailable`
+	f(cfgStr, requestURL, backendHandler, responseExpected)
+	if n := retries.Load(); n != 2 {
+		t.Fatalf("unexpected number of retries; got %d; want 2", n)
+	}
+
+	// retry_status_codes success
+	retries.Store(0)
+	cfgStr = `
+unauthorized_user:
+  url_prefix: ['{BACKEND}/path1', '{BACKEND}/path2']
+  retry_status_codes: [500, 502]`
+	requestURL = "http://some-host.com/foo/?de=fg"
+	backendHandler = func(w http.ResponseWriter, r *http.Request) {
+		if n := retries.Add(1); n < 2 {
+			w.WriteHeader(500)
+			return
+		}
+		fmt.Fprintf(w, "requested_url=http://%s%s", r.Host, r.URL)
+	}
+	responseExpected = `
+statusCode=200
+requested_url={BACKEND}/path2/foo/?de=fg`
+	f(cfgStr, requestURL, backendHandler, responseExpected)
+	if n := retries.Load(); n != 2 {
+		t.Fatalf("unexpected number of retries; got %d; want 2", n)
+	}
 }
 
 type fakeResponseWriter struct {